refactor: 웹소켓 메시지 Enum 클래스 위치 변경
